Added PaymentRouter contract
Added the PaymentRouter.sol contract. Review this contract and see if everything looks good. Please delete/rewrite any/all unnecessary comments that help clean up the flow (there's a ton of them), I'm including them so you can follow along and understand everything as good and fast as possible when reading this for the first time. I ran a Truffle migration test on the core functions we need for MVP, and they appear to function as intended. I can share this migration file if desired, but we should write up some Hardhat tests next and give the contract a more thorough testing of all its functions. The first comments section in the beginning elaborates on what this contract does and what needs to be done before it is ready for production (not much). This is the core of our payment method inside this project, and it is not upgradable, so it is very important that we get this contract right before production.