Uploaded Marketplace and MarketplaceGetters
I'm uploading a whole new file since the changes were pretty drastic and modifying the committed file makes a mess that is difficult to follow. This may be our production contract for the Marketplace (V1.0.0, when deployed), if I don't get time to build out V0.3.0. This takes most of the logic and design from the initial MarketplaceV0 contract and expands it. See the comments section for a detailed description of what has been changed in 0.2.0 and what was fixed in 0.2.1. Also accept changes to PaymentRouter V0.1.1, there is a required bug fix that allows PaymentRouter and Marketplace to play nice together.