Added Testnet Files
I'm going to deploy these files to the Avalanche testnet so front-end can have something to work with. Look over them and see if you can spot any inefficiencies or better design ideas.

Notes:
FactoryPazariTokenMVP: This is a temporary file for testnet only. It doesn't record any information about the contract creation event, and is a standalone file. Use this to test cloning a contract on a live testnet, then find the cloned contract and use it to create market items. I'm redesigning the whole ContractFactory since there is no way we will be able to fit all the contract factories inside the current design. Instead, each factory will be deployed as a standalone contract, and its address will be manually added to FactoryStorage as an approved address that can record data. They will call a storeData() function, or something. However, for now let's just deploy this factory and use it for our first round of testing.

I created interfaces for every external contract, and most of the comments there are going to be used in the documentation.

2_contractTesting_migration: This is just a tester file that I run after every modification to check for breaking changes. It isn't intended for deploying to any live blockchains, just a simple mechanism for doing quick-testing. I'm including it so you can run these basic function tests yourself, but also in case you want to build some better tests in Truffle or Hardhat using the migration tests as a starting point. This test assumes the token contract is already deployed, and picks up from there to mint payment tokens, mint item tokens, create a market item, buy the market item, and then collect payment. It tests a few configurations for the functions as well, but nothing extensive. Most of the peripheral functions like adding and subtracting stock are untested.

PaymentRouter: I made some core design changes to the pull function that makes this contract run much lighter and requires less on-chain storage, and it's also a lot easier to follow now. Now, all pulled earnings are pooled into a single balance mapping for every recipient address, whereas before it was connected to each PaymentRoute and would have required collecting from each route one by one if receiving commissions from more than one PaymentRoute. It's a pretty dramatic UX boost and a huge gas saver for the recipient. It was a pretty bad design flaw on my part, but it's much better now. I also went ahead and implemented a design idea where a recipient's balance mapping defaults to 1 instead of 0, but they still withdraw the tokens they are owed.

I'm sure there are some other things, but those are the biggest I could think of. I just want to get these up here for you.