Final Big Update for PaymentRouter
This is my final "big" update for PaymentRouter. The only thing left to improve is optimizing structs and function inputs for gas efficiency. With these changes PR is now more sustainable for a long-term project, since we now have access control storage and management for this contract. I've implemented AccessControl across all contracts and have run tests to make sure it all works. If one of the AccessControl contracts works, then they should all work, since they are all designed the same or similar.

Each contract has its own localized AccessControl, which can have different admins if we need to assign non-overlapping roles for people. Marketplace admins would be responsible for managing the blacklist and MarketItems, PaymentRouter admins would be responsible for managing route taxes and updating the address for the treasury contract--meanwhile, PazariMVP admins have the sole purpose of recovering lost NFTs sent by accident to the PazariMVP contract. So, we have options now for expanding our team to manage the different components of Pazari.

AccessControl contracts:
In order to use helper contracts we are forced to use tx.origin for msg.sender when the caller is using a Pazari helper/admin contract, but just using tx.origin will expose people to phishing attacks, so I filled out _msgSender()'s logic to discern between Pazari admin contracts (whose msg.sender needs to be invisible and permit tx.origin to go through) and non-admin contracts (whose msg.sender needs to be visible and block out tx.origin) and included it in every contract's AccessControl parent contract. We can manage access control storage now with addAdmin() and removeAdmin() functions. 

Also, PaymentRoutes now have route admins *and* a route creator, and they can be added or removed as well--but the route's original creator will never lose control of the PaymentRoute in the event of a hack (I did the same thing for Marketplace as well, the original creator of a MarketItem *never* loses admin status). I figured the worst thing that could happen to a seller on Pazari is someone compromises their wallet and tries to take over their Pazari items, so I made sure that cannot happen. 

The AccessControl contracts can be run and tested in isolation, and it seems all the "bugs" I ran into yesterday and the day before came from me not writing the Truffle tests correctly. I also included a "memo" parameter for all admin functions that will be reported in the events that are fired from functions, and it would be a good practice if we wrote a quick memo when adding or removing admins so the DB that listens for events will be able to record some context for why these actions were taken.

  PATCH NOTES: MAJOR CHANGES
  - Added AccessControlPR contract, which replaces all access control functions with a single uniform contract that is easier to find and follow.
  - Replaced function modifiers with AccessControlPR modifiers
  - Implemented a design where only Pazari-owned addresses can use the core functions of the PaymentRouter contract.
  -  - This design uses tx.origin and msg.sender combined to detect when a caller is calling from a Pazari contract versus a wallet. If they are calling from a Pazari contract, then tx.origin is used for _msgSender() since we know we can trust our own contracts, but in every other case msg.sender is used (if a phishing attack is attempted, PR will use the phishing contract's address for _msgSender() instead). As long as we repeat this design through every contract it won't be vulnerable to phishing attacks, though it won't prevent someone from getting tricked into using a non-admin contract to create market items so that the hacker's contract becomes the owner of the items. If it becomes a problem, then we can use _msgSender() on the front-end to check the caller's address, since _msgSender() will now reveal intermediate contracts that don't belong to us.
  - Fixed a bug in holdTokens: PaymentRoute was being stored via paymentRouteID mapping inside the function, which for whatever reason wasn't returning the recipients and commissions arrays. Strangely, it works when they're being called by a dedicated helper function. So I created a getter function to address this, and it seems to have been the fix. This was causing all payments to go to one account, even though the logic was correct for iterating through the recipients and commissions. Now, holdTokens and pullTokens work as intended, while never resetting back to 0, thus saving buyers on gas costs when buying items that have multi-recipient payment routes.
 - Implemented design where maxTax and minTax are the absolute upper and lower bounds a routeTax can fall within. Custom taxTypes can no longer assign routeTaxes greater than maxTax. MaxTax will be the absolute max that we permit people to pay, though I'm sure most won't care to pay more than minTax (at least not until there are serious benefits for doing so).

PATCH NOTES: MINOR
 - Changed visibility for tokenBalanceToCollect to internal, is now accessed via a getter function that reports the user's balance adjusted for gas-saving design
 - Did some general re-arranging, re-wrote a few comments, changed a few variable names for presentation, and overall tidying up. I know it looks like I made some big changes, but I really didn't, it's mostly simple cosmetic things, and the most drastic change is the AccessControlPR contract.
 - Added events to constructor. When this contract is deployed it should fire a bunch of events. Not sure if that's very useful or not, but it's there. They will all have the memo "Deployment"
 - Added maxRecipients state variable, which is adjusted by setMaxRecipients(). All new PaymentRoutes must conform to maxRecipients, but changing maxRecipients won't affect pre-existing routes (nor should it). This value is set to 10 in the constructor.
 - Wrapped all ERC20 calls in require() checks where possible so we can use the return values. This was a problem exposed by Slither.

  Comments:
  1. You mentioned the balances were off with the balance +1 design, but my Truffle migration test shows it works fine. One thing to keep in mind is that the treasury's tax is removed *first* before commissions are distributed, so all actual received payments are smaller than the actual price of the item. Thus, at a routeTax of 3% if you send 100 tokens down the PR then your recipients will receive 97 of them (for example, a 50% commission of a $100 item will receive $48.50), and the other 3 tokens go to the treasury. My tests show that this functionality works as intended. All my Remix testing shows that this design should save around 40% on gas costs for buyers using the holdTokens() function, which isn't appreciated as much for single-recipient items (it's like a $0.02 difference, so meh) but should be substantial for multi-recipient items (multiply that $0.02 cents by the number of recipients and then it starts making sense why this design is important). This won't be important at all for pushTokens(), but whereas holdTokens() is the "gas efficient" alternative it makes sense to implement a gas saving design so we can reasonably fit more recipients inside a payment route before needing a specialized contract to handle large numbers of recipients.