Patched Marketplace migration test
Needed to call modifyMarketItem around line 500, and because it wasn't called itemID 1 was never updated and kept its initial itemLimit of 1, which was causing the revert error to throw. This means the contract's logic works, and the test was bugged. I added some logs to show the buyer's balances before and after the final purchases to show that the non-reverting logic works (if buyer's balance + amount >= itemLimit OR if amount == 0 then amount is adjusted to fill the buyer's wallet to the itemLimit).