Security upgrades
I discovered today that I can perform reentrancy calls from the IERC1155 safeTransferFrom function if I fork the code and slip a call to any Marketplace function inside the safeTransferFrom function. I don't know yet if the same hack could be used to turn balanceOf into an attack function, I don't have time to explore the idea and Google doesn't turn up much, but it could be pretty bad if it is possible. If so, then check-effects-interactions would be impossible for this contract, and we would have to use reentrancy guards for security.

So, I took a design from OpenZeppelin to create the noReentrantCalls modifier and placed it in every function that makes external calls to IERC1155. Now, reentrancy attempts are reverted.

I also found the assert check at the end of buyMarketItem() will not throw any errors if the buyer never *actually* receives their token. This would happen if a hacker forked the ERC1155 code to NOT transfer to the buyer when coming from Marketplace, and Marketplace would not catch it. To address this I added a pair of asserts and a pair of local variables to each function that manages inventory. The variables store initial balances, and the asserts check that final balances are where they're supposed to be. 

I noticed at least one assert that expected the balanceOf for Marketplace to be the same as MarketItem.amount, but this is not always true. If a token is directly transferred to Marketplace then balanceOf will be higher than MarketItem.amount, but never the other way around. I removed this check and replaced it with the aforementioned assert check pair, so we will just check initial and final balances instead.

I did have to remove getMarketItemsForSale() to fit these changes, but we can achieve the same result by calling getItemIDsForSale() and feeding the returned array into getMarketItems(), so this function is unnecessary. Just in case we still need it though I'm leaving it commented out, but I doubt we will.

Finally, I overhauled recoverNFT() so it takes in a recipient's address and an admin's memo as parameters, and properly accounts for Marketplace-listed stock, and it also uses better logic that is much easier to understand and follow. It also now emits an admin's event.

Alright, I think this one is for real the last pull request I'll do for hackathon on smart contracts, I want to get everything consolidated in the next few days and finalize it all. We're almost there!