Final touches
These are mostly non-breaking touches, except for one. The one breaking change is recover NFT(), which now takes identical input parameters and emits an identical event as Marketplace. Also added NFTRecovered event, which is identical to NFTRecovered from Marketplace.

The most important touch-up was a patch for the constructor so it can finish executing. Can't read immutable variables before constructor is done, so I just switched out the contract variables with the addresses that are provided to the constructor. I fixed this in main, but not here.

Some formatting improvements, made comments more useful.

The newTokenListing function will now tell you which contract is missing isAdmin for PazariMVP. Should be more helpful in case I can't be reached when we deploy.

Switched out all instances of "memory _memo" with "calldata _memo". I'm doing this across all contracts since it's a harmless optimization, and strings are expensive.